requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic-settings==2.5.0
python-dotenv==1.0.1
google-generativeai==0.8.0
python-multipart==0.0.9
scipy>=1.11.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
